fix(auth): refactor folder structure
