+++ src/application/useCases/user/auth/logout.usecase.ts
@@ -1,9 +1,9 @@
 import { injectable, inject } from "tsyringe";
-import { ILogoutUseCase } from "../../../domain/interfaces/usecaseInterface/auth/logout.usecase.interface";
-import { ITokenService } from "../../../domain/interfaces/serviceInterface/security/token.service.interface";
-import { BaseError } from "../../../domain/errors/base.error";
-import { HTTP_STATUS } from "../../../shared/constants/status-codes";
-import { SUCCESS_MSG } from "../../../shared/constants/success-msg";
+import { ILogoutUseCase } from "../../../../domain/interfaces/usecaseInterface/user/auth/logout.usecase.interface";
+import { ITokenService } from "../../../../domain/interfaces/serviceInterface/security/token.service.interface";
+import { BaseError } from "../../../../domain/errors/base.error";
+import { HTTP_STATUS } from "../../../../shared/constants/status-codes";
+import { SUCCESS_MSG } from "../../../../shared/constants/success-msg";
 
 @injectable()
 export class LogoutUseCase implements ILogoutUseCase {
