+++ src/domain/interfaces/usecaseInterface/user/auth/github-oauth.usecase.interface.ts
@@ -1,4 +1,4 @@
-import { IUserEntity } from "../../../entities/models/user.entity";
+import { IUserEntity } from "../../../../entities/models/user.entity";
 
 export interface IGitHubOAuthLoginUseCase {
   execute(code: string): Promise<{
