+++ src/domain/interfaces/usecaseInterface/user/auth/send-email.usecase.interface.ts
@@ -1,4 +1,4 @@
-import { SendEmailDTO } from "../../../../shared/validation/schemas";
+import { SendEmailDTO } from "../../../../../shared/validation/schemas";
 
 export interface ISendEmailUseCase {
   execute(
