+++ src/infrastructure/di/resolver.ts
@@ -1,24 +1,24 @@
 // src/infrastructure/di/resolver.ts
 import { container } from "tsyringe";
-import { SignupController } from "../../presentation/controllers/auth/signup.controller";
-import { SendOtpController } from "../../presentation/controllers/auth/send-otp.controller";
-import { ResetPasswordController } from "../../presentation/controllers/auth/reset-password.controller";
+import { SignupController } from "../../presentation/controllers/user/auth/signup.controller";
+import { SendOtpController } from "../../presentation/controllers/user/auth/send-otp.controller";
+import { ResetPasswordController } from "../../presentation/controllers/user/auth/reset-password.controller";
 
 import "./container";
 import "./service.registry";
-import { LoginController } from "../../presentation/controllers/auth/login.controller";
-import { RefreshTokenController } from "../../presentation/controllers/auth/refresh-token.controller";
-import { LogoutController } from "../../presentation/controllers/auth/logout.controller";
+import { LoginController } from "../../presentation/controllers/user/auth/login.controller";
+import { RefreshTokenController } from "../../presentation/controllers/user/auth/refresh-token.controller";
+import { LogoutController } from "../../presentation/controllers/user/auth/logout.controller";
 import { AuthMiddleware } from "../../presentation/middleware/auth.middleware";
 import { GetUserController } from "../../presentation/controllers/user/get-user.controller";
-import { ForgotPasswordController } from "../../presentation/controllers/auth/forgot-password.controller";
+import { ForgotPasswordController } from "../../presentation/controllers/user/auth/forgot-password.controller";
 import { AdminSigninController } from "../../presentation/controllers/admin/signin.controller";
 import { AdminMiddleware } from "../../presentation/middleware/admin.middleware";
 import { AdminLogoutController } from "../../presentation/controllers/admin/admin-logout.controller";
 import { GetAllUsersController } from "../../presentation/controllers/admin/get-all-users.controller";
 import { UpdateUserController } from "../../presentation/controllers/admin/update-user.controller";
-import { GoogleOAuthController } from "../../presentation/controllers/auth/google-oauth.controller";
-import { IGitHubOAuthController } from "../../domain/interfaces/controllerInterface/auth/github-oauth.controller.interface";
+import { GoogleOAuthController } from "../../presentation/controllers/user/auth/google-oauth.controller";
+import { IGitHubOAuthController } from "../../domain/interfaces/controllerInterface/user/auth/github-oauth.controller.interface";
 
 // ----------------------------- auth -----------------------------
 export const signupController =
