+++ src/infrastructure/di/usecase.registry.ts
@@ -1,20 +1,20 @@
 import { container } from "tsyringe";
-import { SignupUseCase } from "../../application/useCases/auth/signup.usecase";
-import { ISignupUseCase } from "../../domain/interfaces/usecaseInterface/auth/signup.usecase.interface";
-import { SendEmailUseCase } from "../../application/useCases/auth/send-email.usecase";
-import { ISendEmailUseCase } from "../../domain/interfaces/usecaseInterface/auth/send-email.usecase.interface";
-import { ResetPasswordUseCase } from "../../application/useCases/auth/reset-password.usecase";
-import { IResetPasswordUseCase } from "../../domain/interfaces/usecaseInterface/auth/reset-password.usecase.interface";
-import { IRefreshTokenUseCase } from "../../domain/interfaces/usecaseInterface/auth/refresh-token.usecase.interface";
-import { RefreshTokenUseCase } from "../../application/useCases/auth/refresh-token.usecase";
-import { LoginUseCase } from "../../application/useCases/auth/login.usecase";
-import { ILoginUseCase } from "../../domain/interfaces/usecaseInterface/auth/login.usecase.interface";
-import { LogoutUseCase } from "../../application/useCases/auth/logout.usecase";
-import { ILogoutUseCase } from "../../domain/interfaces/usecaseInterface/auth/logout.usecase.interface";
+import { SignupUseCase } from "../../application/useCases/user/auth/signup.usecase";
+import { ISignupUseCase } from "../../domain/interfaces/usecaseInterface/user/auth/signup.usecase.interface";
+import { SendEmailUseCase } from "../../application/useCases/user/auth/send-email.usecase";
+import { ISendEmailUseCase } from "../../domain/interfaces/usecaseInterface/user/auth/send-email.usecase.interface";
+import { ResetPasswordUseCase } from "../../application/useCases/user/auth/reset-password.usecase";
+import { IResetPasswordUseCase } from "../../domain/interfaces/usecaseInterface/user/auth/reset-password.usecase.interface";
+import { IRefreshTokenUseCase } from "../../domain/interfaces/usecaseInterface/user/auth/refresh-token.usecase.interface";
+import { RefreshTokenUseCase } from "../../application/useCases/user/auth/refresh-token.usecase";
+import { LoginUseCase } from "../../application/useCases/user/auth/login.usecase";
+import { ILoginUseCase } from "../../domain/interfaces/usecaseInterface/user/auth/login.usecase.interface";
+import { LogoutUseCase } from "../../application/useCases/user/auth/logout.usecase";
+import { ILogoutUseCase } from "../../domain/interfaces/usecaseInterface/user/auth/logout.usecase.interface";
 import { GetUserUseCase } from "../../application/useCases/user/get-user.usecase";
 import { IGetUserUseCase } from "../../domain/interfaces/usecaseInterface/user/get-user.usecase.interface";
-import { IForgotPasswordUseCase } from "../../domain/interfaces/usecaseInterface/auth/forgot-password.usecase.interface";
-import { ForgotPasswordUseCase } from "../../application/useCases/auth/forgot-password.usecase";
+import { IForgotPasswordUseCase } from "../../domain/interfaces/usecaseInterface/user/auth/forgot-password.usecase.interface";
+import { ForgotPasswordUseCase } from "../../application/useCases/user/auth/forgot-password.usecase";
 import { IAdminSigninUseCase } from "../../domain/interfaces/usecaseInterface/admin/signin.usecase.interface";
 import { AdminSigninUseCase } from "../../application/useCases/admin/signin.usecase";
 import { IAdminLogoutUseCase } from "../../domain/interfaces/usecaseInterface/admin/admin-logout.usecase.interface";
@@ -23,10 +23,10 @@ import { IGetAllUsersUsecase } from "../../domain/interfaces/usecaseInterface/ad
 import { GetAllUsersUsecase } from "../../application/useCases/admin/get-all-users.usecase";
 import { IUpdateUserUsecase } from "../../domain/interfaces/usecaseInterface/admin/update-user.usecase.interface";
 import { UpdateUserUsecase } from "../../application/useCases/admin/update-user.usecase";
-import { IGoogleOAuthLoginUseCase } from "../../domain/interfaces/usecaseInterface/auth/google-oauth.usecase.interface";
-import { GoogleOAuthLoginUseCase } from "../../application/useCases/auth/google-oauth.usecase";
-import { IGitHubOAuthLoginUseCase } from "../../domain/interfaces/usecaseInterface/auth/github-oauth.usecase.interface";
-import { GitHubOAuthLoginUseCase } from "../../application/useCases/auth/github-oauth.usecase";
+import { IGoogleOAuthLoginUseCase } from "../../domain/interfaces/usecaseInterface/user/auth/google-oauth.usecase.interface";
+import { GoogleOAuthLoginUseCase } from "../../application/useCases/user/auth/google-oauth.usecase";
+import { IGitHubOAuthLoginUseCase } from "../../domain/interfaces/usecaseInterface/user/auth/github-oauth.usecase.interface";
+import { GitHubOAuthLoginUseCase } from "../../application/useCases/user/auth/github-oauth.usecase";
 
 container.register<ISignupUseCase>("ISignupUseCase", {
   useClass: SignupUseCase,
