+++ src/presentation/controllers/user/auth/send-otp.controller.ts
@@ -1,10 +1,10 @@
 import { injectable, inject } from "tsyringe";
 import { Request, Response, NextFunction } from "express";
-import { ISendEmailUseCase } from "../../../domain/interfaces/usecaseInterface/auth/send-email.usecase.interface";
-import { ISendOtpController } from "../../../domain/interfaces/controllerInterface/auth/send-otp.controller.interface";
-import { BaseError } from "../../../domain/errors/base.error";
-import { HTTP_STATUS } from "../../../shared/constants/status-codes";
-import { ERROR_MSG } from "../../../shared/constants/error-msg";
+import { ISendEmailUseCase } from "../../../../domain/interfaces/usecaseInterface/user/auth/send-email.usecase.interface";
+import { ISendOtpController } from "../../../../domain/interfaces/controllerInterface/user/auth/send-otp.controller.interface";
+import { BaseError } from "../../../../domain/errors/base.error";
+import { HTTP_STATUS } from "../../../../shared/constants/status-codes";
+import { ERROR_MSG } from "../../../../shared/constants/error-msg";
 
 @injectable()
 export class SendOtpController implements ISendOtpController {
